Change description of choclo_arepa.html
As I changed all dishes to new ones, it made no sense to talk about
choclo arepa as the only one that comes from Colombia (given that now
all of them do). Also, now the description has a higher level of detail
and explains what it is and why is it different to other arepas.